Handle issue with InterruptableThread
Sometimes InterruptableThread can't be interrupted because get_id()
returns None (seems to depend on logging call from stage, so this is
probably a race condition?). Can't track it down right now, so for now
it prints a warning and just fails to interrupt the thread. Shouldn't be
an issue for the real stage anyway, as that won't use time.sleep().